+++ db-queue-scheduler-core/src/main/java/ru/yoomoney/tech/dbqueue/scheduler/config/DatabaseDialect.java
@@ -12,6 +12,17 @@ public enum DatabaseDialect {
      */
     POSTGRESQL,
     /**
+     * Microsoft SQL Server
+     */
+    MSSQL,
+    /**
+     * Oracle 11g
+     *
+     * <p> This version doesn't have automatically incremented primary keys,
+     * so you must specify sequence name in a scheduler configurator.
+     */
+    ORACLE_11G,
+    /**
      * H2 in-memory database
      */
     H2