+++ db-queue-scheduler-spring/src/main/java/ru/yoomoney/tech/dbqueue/scheduler/SpringSchedulerConfigurator.java
@@ -14,6 +14,7 @@ import ru.yoomoney.tech.dbqueue.scheduler.internal.schedule.NextExecutionTimePro
 import ru.yoomoney.tech.dbqueue.spring.dao.SpringDatabaseAccessLayer;
 
 import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
 
 import static java.util.Objects.requireNonNull;
 
@@ -40,6 +41,7 @@ import static java.util.Objects.requireNonNull;
  */
 public class SpringSchedulerConfigurator implements SchedulerConfigurator {
     private String tableName;
+    private String idSequenceName;
     private DatabaseDialect databaseDialect;
     private JdbcOperations jdbcOperations;
     private TransactionOperations transactionOperations;
@@ -60,7 +62,7 @@ public class SpringSchedulerConfigurator implements SchedulerConfigurator {
      *     reenqueue_attempt INTEGER                  DEFAULT 0,
      *     total_attempt     INTEGER                  DEFAULT 0
      *  );
-     *  CREATE UNIQUE INDEX scheduled_tasks_name_queue_name_uq ON scheduled_tasks (queue_name);
+     *  CREATE UNIQUE INDEX scheduled_tasks_uq ON scheduled_tasks (queue_name);
      *  }</pre>
      *
      * @param tableName table name that stores scheduled tasks
@@ -69,6 +71,17 @@ public class SpringSchedulerConfigurator implements SchedulerConfigurator {
     public SpringSchedulerConfigurator withTableName(@Nonnull String tableName) {
         requireNonNull(tableName, "tableName");
         this.tableName = tableName;
+        return this;
+    }
+
+    /**
+     * Sets sequence name for generating primary key of tasks table.
+     *
+     * @param idSequenceName sequence name for generating primary key of tasks table.
+     * @return the same instance of {@link SpringSchedulerConfigurator}
+     */
+    public SpringSchedulerConfigurator withIdSequenceName(@Nullable String idSequenceName) {
+        this.idSequenceName = idSequenceName;
         return this;
     }
 
@@ -139,6 +152,7 @@ public class SpringSchedulerConfigurator implements SchedulerConfigurator {
         );
         ScheduledTaskQueueDao scheduledTaskQueueDao = new DefaultScheduledTaskQueueDao(
                 tableName,
+                databaseDialect,
                 jdbcOperations,
                 transactionOperations,
                 QueueTableSchema.builder().build()
@@ -146,6 +160,7 @@ public class SpringSchedulerConfigurator implements SchedulerConfigurator {
         return new DefaultScheduler(
                 new ScheduledTaskManagerBuilder()
                         .withTableName(tableName)
+                        .withIdSequenceName(idSequenceName)
                         .withScheduledTaskQueueDao(scheduledTaskQueueDao)
                         .withDatabaseAccessLayer(databaseAccessLayer)
                         .withScheduledTaskLifecycleListener(scheduledTaskLifecycleListener)
@@ -160,6 +175,10 @@ public class SpringSchedulerConfigurator implements SchedulerConfigurator {
                 return ru.yoomoney.tech.dbqueue.config.DatabaseDialect.POSTGRESQL;
             case H2:
                 return ru.yoomoney.tech.dbqueue.config.DatabaseDialect.H2;
+            case ORACLE_11G:
+                return ru.yoomoney.tech.dbqueue.config.DatabaseDialect.ORACLE_11G;
+            case MSSQL:
+                return ru.yoomoney.tech.dbqueue.config.DatabaseDialect.MSSQL;
             default:
                 throw new IllegalStateException("got unsupported databaseDialect: databaseDialect=" + databaseDialect);
         }