+++ db-queue-scheduler-spring/src/main/java/ru/yoomoney/tech/dbqueue/scheduler/internal/db/DefaultScheduledTaskQueueDao.java
@@ -4,14 +4,17 @@ import org.springframework.jdbc.core.JdbcOperations;
 import org.springframework.jdbc.core.namedparam.NamedParameterJdbcTemplate;
 import org.springframework.transaction.support.TransactionOperations;
 import ru.yoomoney.tech.dbqueue.config.QueueTableSchema;
+import ru.yoomoney.tech.dbqueue.scheduler.config.DatabaseDialect;
 import ru.yoomoney.tech.dbqueue.settings.QueueId;
 
 import javax.annotation.Nonnull;
 import java.sql.Timestamp;
 import java.time.Instant;
+import java.util.Calendar;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
+import java.util.TimeZone;
 
 import static java.util.Objects.requireNonNull;
 
@@ -26,20 +29,24 @@ import static java.util.Objects.requireNonNull;
 public class DefaultScheduledTaskQueueDao implements ScheduledTaskQueueDao {
 
     private final String tableName;
+    private final DatabaseDialect databaseDialect;
     private final QueueTableSchema queueTableSchema;
     private final TransactionOperations transactionOperations;
     private final NamedParameterJdbcTemplate namedParameterJdbcTemplate;
 
     public DefaultScheduledTaskQueueDao(@Nonnull String tableName,
+                                        @Nonnull DatabaseDialect databaseDialect,
                                         @Nonnull JdbcOperations jdbcOperations,
                                         @Nonnull TransactionOperations transactionOperations,
                                         @Nonnull QueueTableSchema queueTableSchema) {
         requireNonNull(tableName, "tableName");
+        requireNonNull(databaseDialect, "databaseDialect");
         requireNonNull(jdbcOperations, "jdbcOperations");
         requireNonNull(transactionOperations, "transactionOperations");
         requireNonNull(queueTableSchema, "queueTableSchema");
 
         this.queueTableSchema = queueTableSchema;
+        this.databaseDialect = databaseDialect;
         this.namedParameterJdbcTemplate = new NamedParameterJdbcTemplate(jdbcOperations);
         this.transactionOperations = transactionOperations;
         this.tableName = tableName;
@@ -78,9 +85,19 @@ public class DefaultScheduledTaskQueueDao implements ScheduledTaskQueueDao {
                 queueTableSchema.getNextProcessAtField(),
                 queueTableSchema.getQueueNameField()
         );
+        Calendar calendar;
+        if (databaseDialect == DatabaseDialect.MSSQL) {
+            // https://github.com/Microsoft/mssql-jdbc/issues/758
+            calendar = Calendar.getInstance(TimeZone.getTimeZone("UTC"));
+            calendar.setTime(Timestamp.from(nextProcessDate));
+        } else {
+            // on the other hand oracle and h2 can't process calendar properly
+            calendar = Calendar.getInstance();
+            calendar.setTime(Timestamp.from(nextProcessDate));
+        }
         Integer updatedRows = transactionOperations.execute(status -> namedParameterJdbcTemplate.update(
                 rescheduleQuery,
-                Map.<String, Object>of("queueName", queueId.asString(), "nextProcessDate", Timestamp.from(nextProcessDate))
+                Map.<String, Object>of("queueName", queueId.asString(), "nextProcessDate", calendar)
         ));
         return updatedRows == null ? 0 : updatedRows;
     }