+++ db-queue-scheduler-spring/src/test/java/ru/yoomoney/tech/dbqueue/scheduler/BaseTest.java
@@ -1,15 +1,16 @@
 package ru.yoomoney.tech.dbqueue.scheduler;
 
-import org.h2.jdbcx.JdbcDataSource;
 import org.junit.jupiter.api.BeforeAll;
-import org.postgresql.ds.PGSimpleDataSource;
-import org.springframework.jdbc.core.JdbcTemplate;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
-import org.springframework.transaction.support.TransactionTemplate;
+import org.testcontainers.containers.MSSQLServerContainer;
+import org.testcontainers.containers.OracleContainer;
 import org.testcontainers.containers.PostgreSQLContainer;
 import org.testcontainers.junit.jupiter.Container;
 import org.testcontainers.junit.jupiter.Testcontainers;
-import ru.yoomoney.tech.dbqueue.scheduler.config.DatabaseDialect;
+import ru.yoomoney.tech.dbqueue.scheduler.db.DatabaseAccess;
+import ru.yoomoney.tech.dbqueue.scheduler.db.H2DatabaseAccessConfigurator;
+import ru.yoomoney.tech.dbqueue.scheduler.db.MsSqlDatabaseAccessConfigurator;
+import ru.yoomoney.tech.dbqueue.scheduler.db.OracleDatabaseAccessConfigurator;
+import ru.yoomoney.tech.dbqueue.scheduler.db.PostgresDatabaseAccessConfigurator;
 
 import java.util.concurrent.atomic.AtomicLong;
 import java.util.stream.Stream;
@@ -22,102 +23,32 @@ import java.util.stream.Stream;
  */
 @Testcontainers
 public abstract class BaseTest {
-    private static final String POSTGRES_SCHEDULED_TASKS_TABLE_DDL = "" +
-            "CREATE TABLE scheduled_tasks (" +
-            "  id                BIGSERIAL PRIMARY KEY," +
-            "  queue_name        TEXT NOT NULL," +
-            "  payload           TEXT," +
-            "  created_at        TIMESTAMP WITH TIME ZONE DEFAULT now()," +
-            "  next_process_at   TIMESTAMP WITH TIME ZONE DEFAULT now()," +
-            "  attempt           INTEGER                  DEFAULT 0," +
-            "  reenqueue_attempt INTEGER                  DEFAULT 0," +
-            "  total_attempt     INTEGER                  DEFAULT 0" +
-            ");" +
-            "CREATE UNIQUE INDEX scheduled_tasks_name_queue_name_uq ON scheduled_tasks (queue_name);";
-
-    private static final String H2_SCHEDULED_TASKS_TABLE_DDL = "" +
-            "CREATE TABLE scheduled_tasks (" +
-            "  id                BIGSERIAL PRIMARY KEY," +
-            "  queue_name        VARCHAR(100) NOT NULL," +
-            "  payload           VARCHAR(100)," +
-            "  created_at        TIMESTAMP WITH TIME ZONE DEFAULT now()," +
-            "  next_process_at   TIMESTAMP WITH TIME ZONE DEFAULT now()," +
-            "  attempt           INTEGER                  DEFAULT 0," +
-            "  reenqueue_attempt INTEGER                  DEFAULT 0," +
-            "  total_attempt     INTEGER                  DEFAULT 0" +
-            ");" +
-            "CREATE UNIQUE INDEX scheduled_tasks_name_queue_name_uq ON scheduled_tasks (queue_name);";
+    @Container
+    private static final MSSQLServerContainer<?> mssqlServerContainer =
+            new MSSQLServerContainer<>("mcr.microsoft.com/mssql/server:2019-CU1-ubuntu-16.04").acceptLicense();
 
     @Container
     private static final PostgreSQLContainer<?> postgresqlContainer = new PostgreSQLContainer<>("postgres:9.5");
 
+    @Container
+    private static final OracleContainer oracleContainer = new OracleContainer("gvenzl/oracle-xe:11.2.0.2-slim");
+
     protected static DatabaseAccess postgres;
+    protected static DatabaseAccess oracle;
+    protected static DatabaseAccess mssql;
     protected static DatabaseAccess h2;
+
     protected static AtomicLong uniqueCounter = new AtomicLong();
 
     @BeforeAll
     public static void configureDatabases() {
-        postgres = configurePostgres();
-        h2 = configureH2();
-    }
-
-    private static DatabaseAccess configurePostgres() {
-        PGSimpleDataSource dataSource = new PGSimpleDataSource();
-        dataSource.setUrl(postgresqlContainer.getJdbcUrl());
-        dataSource.setPassword(postgresqlContainer.getPassword());
-        dataSource.setUser(postgresqlContainer.getUsername());
-
-        JdbcTemplate jdbcTemplate = new JdbcTemplate(dataSource);
-        TransactionTemplate transactionTemplate = new TransactionTemplate(new DataSourceTransactionManager(dataSource));
-        transactionTemplate.executeWithoutResult(status -> jdbcTemplate.update(POSTGRES_SCHEDULED_TASKS_TABLE_DDL));
-        return new DatabaseAccess(DatabaseDialect.POSTGRESQL, jdbcTemplate, transactionTemplate);
-    }
-
-    private static DatabaseAccess configureH2() {
-        JdbcDataSource dataSource = new JdbcDataSource();
-        dataSource.setURL("jdbc:h2:~/scheduled_task_h2");
-        dataSource.setUser("sa");
-        dataSource.setPassword("sa");
-
-        JdbcTemplate jdbcTemplate = new JdbcTemplate(dataSource);
-        TransactionTemplate transactionTemplate = new TransactionTemplate(new DataSourceTransactionManager(dataSource));
-        transactionTemplate.executeWithoutResult(status -> jdbcTemplate.update("DROP ALL OBJECTS DELETE FILES"));
-        transactionTemplate.executeWithoutResult(status -> jdbcTemplate.update(H2_SCHEDULED_TASKS_TABLE_DDL));
-        return new DatabaseAccess(DatabaseDialect.H2, jdbcTemplate, transactionTemplate);
+        postgres = PostgresDatabaseAccessConfigurator.configure(postgresqlContainer);
+        oracle = OracleDatabaseAccessConfigurator.configure(oracleContainer);
+        mssql = MsSqlDatabaseAccessConfigurator.configure(mssqlServerContainer);
+        h2 = H2DatabaseAccessConfigurator.configure();
     }
 
     public static Stream<DatabaseAccess> databaseAccessStream() {
-        return Stream.of(postgres, h2);
-    }
-
-    protected static class DatabaseAccess {
-        private final DatabaseDialect databaseDialect;
-        private final JdbcTemplate jdbcTemplate;
-        private final TransactionTemplate transactionTemplate;
-
-        DatabaseAccess(DatabaseDialect databaseDialect, JdbcTemplate jdbcTemplate, TransactionTemplate transactionTemplate) {
-            this.databaseDialect = databaseDialect;
-            this.jdbcTemplate = jdbcTemplate;
-            this.transactionTemplate = transactionTemplate;
-        }
-
-        public DatabaseDialect getDatabaseDialect() {
-            return databaseDialect;
-        }
-
-        public JdbcTemplate getJdbcTemplate() {
-            return jdbcTemplate;
-        }
-
-        public TransactionTemplate getTransactionTemplate() {
-            return transactionTemplate;
-        }
-
-        @Override
-        public String toString() {
-            return "DatabaseAccess{" +
-                    "databaseDialect=" + databaseDialect +
-                    '}';
-        }
+        return Stream.of(postgres, oracle, mssql, h2);
     }
 }